added explicit Dialog Component
