+++ src/components/Dialog/Test/Dialog.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { boolean, object, text, withKnobs } from '@storybook/addon-knobs';
-import DialogComponent from '../Dialog.Component';
+import DialogComponent from '../Dialog Story';
 import { action } from '@storybook/addon-actions';
 import { storiesOf } from '@storybook/react';
 import  DialogHeader from '../DialogHeader.Component';
